+++ src/onegov.activity/onegov/activity/models/period.py
@@ -61,6 +61,9 @@ class Period(Base, TimestampMixin):
     #: or for each single booking
     all_inclusive = Column(Boolean, nullable=False, default=False)
 
+    #: Time between bookings in minutes
+    minutes_between = Column(Integer, nullable=True, default=0)
+
     __table_args__ = (
         CheckConstraint((
             '"prebooking_start" <= "prebooking_end" AND '